Tighten CRD validation naming and drop dead production helper
This follow-up only addresses review-driven cleanup: it renames the CRD client variable to a more direct name and removes the unused production-side checkRequiredFields helper after moving the tiny helper logic into the test file. The runtime behavior stays the same; the change just makes the manifest-driven implementation clearer and reduces dead code in the production path.

Constraint: Keep this follow-up scoped to naming and dead-code cleanup without changing validation semantics
Rejected: Leave the helper in production code because tests still use it | obscures the real runtime path and keeps dead logic around
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep test-only schema helpers in test files unless they are part of the actual runtime algorithm
Tested: gofmt; go test ./pkg/kube -run 'TestFieldExistsInSchema|TestCheckRequiredFields|TestRequiredCRDsDefinition|TestCheckCRDVersionsWithClient_AllValid|TestCheckCRDVersionsWithClient_StorageVersionMismatch|TestCheckCRDVersionsWithClient_ListError|TestCheckCRDVersionsWithClient_MissingManifestField|TestCheckCRDVersionsWithClient_OptionalPathBypass'; git diff --check
Not-tested: Full repository test matrix
EOF && git -C /Users/aias/Work/github/aias00/higress-pr-3355 push aias00 HEAD:feat/crd_version_check
